move quota
