+++ ydb/core/persqueue/pqtablet/partition/read_quoter.cpp
@@ -1,5 +1,4 @@
 #include "read_quoter.h"
-#include "account_read_quoter.h"
 
 #include <ydb/core/persqueue/public/constants.h>
