+++ ydb/core/persqueue/pqtablet/partition/user_info.h
@@ -1,10 +1,11 @@
 #pragma once
 
+#include "account_read_quoter.h"
+
 #include <ydb/core/persqueue/working_time_counter.h>
 #include <ydb/core/persqueue/subscriber.h>
 #include <ydb/core/persqueue/percentile_counter.h>
-#include <ydb/core/persqueue/quota_tracker.h>
-#include <ydb/core/persqueue/account_read_quoter.h>
+#include "quota_tracker.h"
 #include <ydb/core/persqueue/public/constants.h>
 #include <ydb/core/persqueue/pqtablet/metering_sink.h>
 #include <ydb/core/persqueue/dread_cache_service/caching_service.h>
